+++ src/main/connector/connector-manager.ts
@@ -143,7 +143,7 @@ export class ConnectorManager {
 
         logger.debug('Updating autocompletion', { baseDir: connector.baseDir });
         if (connector.taskId) {
-          void this.projectManager.getProject(connector.baseDir).updateAutocompletionData(connector.taskId, message.words, message.models);
+          void this.projectManager.getProject(connector.baseDir).updateAutocompletionData(connector.taskId, message.words);
         } else {
           // Handle the case where taskId is not available, e.g., log an error or apply to all tasks
           logger.warn('Received update-autocompletion message from a connector without a taskId.');